agent/requirements.txt
# Core LLM & Orchestration
rich>=13.0.0
pyyaml>=6.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.2.50,<0.3
langgraph-checkpoint>=2.0.0
python-dotenv>=1.0.0

# Data & Computation
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
duckdb>=1.2.0

# ML Strategy
scikit-learn>=1.3.0
joblib>=1.3.0

# Signal Skills
smartmoneyconcepts>=0.0.1
pyharmonics>=1.5.0

# Data Providers
tushare>=1.2.89
requests>=2.31.0
yfinance>=0.2.30

# REST API & SSE
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# MCP Server
fastmcp>=2.0.0
